moving algorithm_service into database package
