+++ src/recordlinker/routes/algorithm_router.py
@@ -12,8 +12,8 @@ import fastapi
 import sqlalchemy.orm as orm
 
 from recordlinker import schemas
+from recordlinker.database import algorithm_service as service
 from recordlinker.database import get_session
-from recordlinker.linking import algorithm_service as service
 
 router = fastapi.APIRouter()
